Proxy beans with primitive constructor parameters.
Now it is possible to proxy beans that take primitive types parameters
in constructors. Before this change, objects that took int, long,
boolean, etc, could not be proxied because the auto proxy creator did
not attempt to convert a string to the primitive value. This patch uses
spring's DataBinder to convert the values in the bean definition to the
expected argument type.